requirements.txt
numpy
opencv-python
pillow
scikit-image
scipy
torch
torchvision
tqdm